Update README.md
The WSDA Music company wants to make a sale plan and needs data-driven insights to make informed decisions.
However, the WSDA Music management team noticed a discrepancy in the company’s finances, which had accrued between 2011 and 2012. 
My task is to gather information about the discrepancy in sales and pinpoint the reason or suspect.